+++ clientes-app/src/app/altapersona/altapersona.component.ts
@@ -10,13 +10,10 @@ export class AltapersonaComponent implements OnInit {
 
 
   personas: Persona[] = [
-    new Persona('Juan','Perez'),
-    new Persona('Laura', 'Juarez'),
-    new Persona('Karla', 'Lara')
+
   ];
 
   constructor() {
-    this.personas[1]= new Persona('Juan','Perez');
    }
 
   ngOnInit(): void {