+++ clientes-app/src/app/empleados/empleado.model.ts
@@ -1,0 +1,9 @@
+import { Persona } from "../persona/persona.model";
+
+export class Empleado{
+
+  persona!: Persona;
+  ccc!: string;
+
+
+}