+++ clientes-app/src/app/empleados/empleados.component.ts
@@ -1,4 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { Usuario } from './../login/usuario.model';
+import { EmpleadoService } from './empleado.service';
+import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-empleados',
@@ -7,9 +9,23 @@ import { Component, OnInit } from '@angular/core';
 })
 export class EmpleadosComponent implements OnInit {
 
-  constructor() { }
+
+  usuario!: Usuario;
+
+  constructor(private empleadoService: EmpleadoService) {
+  }
 
   ngOnInit(): void {
+    this.empleadoService.obtenerUsuarioPorUsername().subscribe(
+      usu => this.usuario = usu
+    );
+  }
+
+  guaradarEmpleado(): void{
+    console.log('entra' + this.usuario)
+    this.empleadoService.guardarEmpleado(this.usuario);
+
+
   }
 
 }