+++ clientes-app/src/app/login/login.component.ts
@@ -1,4 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+import { PersonaServices } from './../persona/persona.service';
+import { Usuario } from './usuario.model';
+import { Component, Input, OnInit } from '@angular/core';
+import { Persona } from '../persona/persona.model';
+import swal from 'sweetalert2';
 
 @Component({
   selector: 'app-login',
@@ -9,9 +14,38 @@ export class LoginComponent implements OnInit {
 
   titulo='Acceso'
 
-  constructor() { }
+  usuario: Usuario;
 
-  ngOnInit(): void {
+  constructor(private personaServices: PersonaServices, private router: Router) {
+    this.usuario=new Usuario();
   }
 
+  ngOnInit(): void {
+    if (this.personaServices.isAuthenticated()) {
+      swal.fire('Login', `Hola ${this.personaServices.usuario.username} ya estás autenticado!`, 'info');
+      this.router.navigate(['/productosgrid']);
+    }
+  }
+
+  login(): void{
+    console.log("usuario:"+this.usuario);
+    if(this.usuario.username== null || this.usuario.password==null){
+        swal.fire('Error login', 'usuario o contraseña vacia', 'error')
+    }
+    this.personaServices.login(this.usuario).subscribe(response =>
+      {
+        this.personaServices.guardarUsuario(response.access_token);
+        this.personaServices.guardarToken(response.access_token);
+        let payload= JSON.parse(atob(response.access_token.split(".")[1]));
+        console.log(payload);
+        this.router.navigate(['productosgrid']);
+      }, error =>{
+        if (error.status==400){
+          swal.fire('Error login', 'usuario o contraseña incorrectos', 'error')
+        }
+
+      }
+
+      );
+  }
 }