+++ clientes-app/src/app/login/usuario.model.ts
@@ -1,0 +1,12 @@
+export class Usuario{
+
+    id:number;
+    nombre: string;
+    apellido: string;
+    edad: number;
+    username: string;
+    password: string;
+    roles: string[]=[];
+
+
+}