+++ clientes-app/src/app/persona/persona.model.ts
@@ -1,11 +1,13 @@
 export class Persona{
 
+    id: number;
     nombre: string = '';
     apellido: string = '';
 
-    constructor(nombre: string, apellido: string){
+    constructor(nombre: string, apellido: string, id:number){
         this.nombre = nombre;
         this.apellido = apellido;
+        this.id=id;
     }