+++ clientes-app/src/app/persona/persona.service.ts
@@ -1,0 +1,109 @@
+import { Usuario } from './../login/usuario.model';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Injectable, OnInit } from "@angular/core";
+import { Observable } from "rxjs";
+import { Persona } from "./persona.model";
+
+@Injectable({
+  providedIn: "root"
+})
+
+export class PersonaServices implements OnInit{
+
+  private _usuario: Usuario;
+  private _token: any;
+  urlToken: string ="http://localhost:8090/api/security/oauth/token";
+  private httpHeaders = new HttpHeaders({ 'Content-Type': 'application/json' });
+
+  constructor(private httpClient: HttpClient){}
+
+
+
+  ngOnInit(): void {
+    throw new Error("Method not implemented.");
+  }
+
+  login(usuario: Usuario):Observable<any>{
+    const urlEndpoint = 'http://localhost:8090/api/security/oauth/token';
+
+    const credenciales = btoa('frontendapp' + ':' + '12345');
+
+    const httpHeaders = new HttpHeaders({
+      'Content-Type': 'application/x-www-form-urlencoded',
+      'Authorization': 'Basic ' + credenciales
+    });
+
+    let params = new URLSearchParams();
+    params.set('grant_type', 'password');
+    params.set('username', usuario.username);
+    params.set('password', usuario.password);
+    console.log(params.toString());
+    return this.httpClient.post<any>(urlEndpoint, params.toString(), { headers: httpHeaders });
+  }
+
+  guardarUsuario(accessToken: string): void {
+    let payload = this.obtenerDatosToken(accessToken);
+    this._usuario = new Usuario();
+    this._usuario.nombre = payload.nombre;
+    this._usuario.apellido = payload.apellido;
+    this._usuario.username = payload.user_name;
+    this._usuario.roles = payload.authorities;
+    sessionStorage.setItem('usuario', JSON.stringify(this._usuario));
+  }
+
+  guardarToken(accessToken: string): void {
+    this._token = accessToken;
+    sessionStorage.setItem('token', accessToken);
+  }
+
+  obtenerDatosToken(accessToken: any): any {
+    if (accessToken != null) {
+      return JSON.parse(atob(accessToken.split(".")[1]));
+    }
+    return null;
+  }
+
+  public get usuario(): Usuario {
+    if (this._usuario != null) {
+      return this._usuario;
+    } else if (this._usuario == null && sessionStorage.getItem('usuario') != null) {
+        this._usuario = JSON.parse(sessionStorage.getItem('usuario')!) as Usuario;
+      return this._usuario;
+    }
+    return new Usuario();
+  }
+
+  public get token(): any {
+    if (this._token != null) {
+      return this._token;
+    } else if (this._token == null && sessionStorage.getItem('token') != null) {
+      this._token = sessionStorage.getItem('token');
+      return this._token;
+    }
+    return null;
+  }
+
+  isAuthenticated(): boolean {
+    let payload = this.obtenerDatosToken(this.token);
+    if (payload != null && payload.user_name && payload.user_name.length > 0) {
+      return true;
+    }
+    return false;
+  }
+
+  logout(): void {
+    this._token = null;
+    this._usuario = new Usuario();
+    sessionStorage.clear();
+    sessionStorage.removeItem('token');
+    sessionStorage.removeItem('usuario');
+  }
+
+  hasRole(role: string): boolean {
+    if (this.usuario.roles.includes(role)) {
+      return true;
+    }
+    return false;
+  }
+
+}