+++ ectemplate-servicio-personas/src/main/java/net/ectemplate/app/personas/services/PersonasServicesI.java
@@ -7,5 +7,7 @@ import net.ectemplate.app.personas.entities.Usuario;
 public interface PersonasServicesI {
 	
 	public List<Usuario> findAll();
+	
+	public Usuario actualizaUsuario (Usuario usuario);
 
 }