+++ ectemplate-servicio-personas/src/main/java/net/ectemplate/app/personas/services/PersonasServicesImpl.java
@@ -21,4 +21,10 @@ public class PersonasServicesImpl implements PersonasServicesI{
 		return (List<Usuario>)personasRepository.findAll();
 	}
 
+	@Override
+	public Usuario actualizaUsuario(Usuario usuario) {
+		return personasRepository.save(usuario);
+	}
+
+	
 }