+++ ectemplate-servicio-personas/src/main/resources/import.sql
@@ -1,5 +1,5 @@
 INSERT INTO usuarios (nombre, apellidos, edad, username, password) VALUES ('nombre1', 'apellidos1', 10, 'usuario1', '$2a$10$5dfOSR0v.wnEk8CHTtXQvudO.dM0/UIPSVFK71yZXojITFbR8Qfp.');
-INSERT INTO usuarios (nombre, apellidos, edad, username, password) VALUES ('nombre2', 'apellidos2', 20, 'usuario2', '$2a$10$ypLY0nLbod0nZAkEpofD.eEzXzGJmcnpx7nxQf7rWnZzBudyEavQC.');
+INSERT INTO usuarios (nombre, apellidos, edad, username, password) VALUES ('nombre2', 'apellidos2', 20, 'usuario2', '$2a$10$5dfOSR0v.wnEk8CHTtXQvudO.dM0/UIPSVFK71yZXojITFbR8Qfp.');
 
 INSERT INTO roles (nombre) VALUES ('ROLE_USER');
 INSERT INTO roles (nombre) VALUES ('ROLE_ADMIN');